+++ R/extraction.R
@@ -1,6 +1,6 @@
 #' @importFrom stats is.ts
 regroup_ts <- function(x) {
-    if (is.ts(x)) {
+    if (stats::is.ts(x)) {
         return(list(x))
     }
     if (is.list(x)) {
@@ -31,7 +31,7 @@ regroup_ts <- function(x) {
 #' - `date`: observation dates,
 #' - `value`: numeric values of the series.
 #'
-#' @examplesIf rjd3toolkit::get_java_version() >= rjd3toolkit::minimal_java_version
+#' @examplesIf rjd3jars::check_java_version(silent = TRUE)
 #'
 #' library("rjd3toolkit")
 #' library("rjd3workspace")
@@ -78,7 +78,7 @@ get_series.JD3_TRAMOSEATS_RSLTS <- function(x, name, ...) {
                 output,
                 data.frame(
                     series = s,
-                    date = series |> time() |> zoo::as.Date(),
+                    date = series |> stats::time() |> zoo::as.Date(),
                     value = as.numeric(series)
                 )
             )
@@ -137,7 +137,7 @@ get_series.jobjRef <- function(x, ...) {
 #'
 #' @returns A Java Seasonal Adjustment Item object (`jsai`).
 #'
-#' @examplesIf rjd3toolkit::get_java_version() >= rjd3toolkit::minimal_java_version
+#' @examplesIf rjd3jars::check_java_version(silent = TRUE)
 #'
 #' library("rjd3toolkit")
 #' library("rjd3workspace")
@@ -177,7 +177,7 @@ get_jsai_by_name <- function(jws, series_name) {
 #'
 #' @returns a list with all the groups and named variables
 #'
-#' @examplesIf rjd3toolkit::get_java_version() >= rjd3toolkit::minimal_java_version
+#' @examplesIf rjd3jars::check_java_version(silent = TRUE)
 #' context_FR <- create_insee_context()
 #' get_named_variables(context_FR)
 #'