+++ R/random-spec.R
@@ -419,7 +419,7 @@ random_add_usrdefvar <- function(x) {
 #'
 #' @returns a JD+ Specification
 #'
-#' @examplesIf rjd3toolkit::get_java_version() >= rjd3toolkit::minimal_java_version
+#' @examplesIf rjd3jars::check_java_version(silent = TRUE)
 #' set.seed(1L)
 #' spec <- random_spec()
 #'