+++ R/rjd3production-package.R
@@ -2,7 +2,6 @@
 "_PACKAGE"
 
 ## usethis namespace: start
-#' @importFrom rjd3workspace read_sai
-#' @importFrom rjd3toolkit get_java_version minimal_java_version
+#' @importFrom rjd3jars check_java_version
 ## usethis namespace: end
 NULL