+++ tests/testthat/test-translation.R
@@ -1,10 +1,8 @@
 set.seed(2026L)
 
 test_that("rev_set_x11 works", {
-    cond_skip_java <- rjd3toolkit::get_java_version() <
-        rjd3toolkit::minimal_java_version
-    testthat::skip_if(
-        condition = cond_skip_java,
+    testthat::skip_if_not(
+        condition = rjd3jars::check_java_version(silent = TRUE),
         message = "Java version is not sufficient."
     )
     for (k in seq_len(100L)) {
@@ -25,10 +23,8 @@ test_that("rev_set_x11 works", {
 })
 
 test_that("rev_add_ramp works", {
-    cond_skip_java <- rjd3toolkit::get_java_version() <
-        rjd3toolkit::minimal_java_version
-    testthat::skip_if(
-        condition = cond_skip_java,
+    testthat::skip_if_not(
+        condition = rjd3jars::check_java_version(silent = TRUE),
         message = "Java version is not sufficient."
     )
     for (k in seq_len(100L)) {
@@ -49,10 +45,8 @@ test_that("rev_add_ramp works", {
 })
 
 test_that("rev_set_transform works", {
-    cond_skip_java <- rjd3toolkit::get_java_version() <
-        rjd3toolkit::minimal_java_version
-    testthat::skip_if(
-        condition = cond_skip_java,
+    testthat::skip_if_not(
+        condition = rjd3jars::check_java_version(silent = TRUE),
         message = "Java version is not sufficient."
     )
     for (k in seq_len(100L)) {
@@ -73,10 +67,8 @@ test_that("rev_set_transform works", {
 })
 
 test_that("rev_set_easter works", {
-    cond_skip_java <- rjd3toolkit::get_java_version() <
-        rjd3toolkit::minimal_java_version
-    testthat::skip_if(
-        condition = cond_skip_java,
+    testthat::skip_if_not(
+        condition = rjd3jars::check_java_version(silent = TRUE),
         message = "Java version is not sufficient."
     )
     for (k in seq_len(100L)) {
@@ -97,10 +89,8 @@ test_that("rev_set_easter works", {
 })
 
 test_that("rev_set_basic works", {
-    cond_skip_java <- rjd3toolkit::get_java_version() <
-        rjd3toolkit::minimal_java_version
-    testthat::skip_if(
-        condition = cond_skip_java,
+    testthat::skip_if_not(
+        condition = rjd3jars::check_java_version(silent = TRUE),
         message = "Java version is not sufficient."
     )
     for (k in seq_len(100L)) {
@@ -121,10 +111,8 @@ test_that("rev_set_basic works", {
 })
 
 test_that("rev_set_estimate works", {
-    cond_skip_java <- rjd3toolkit::get_java_version() <
-        rjd3toolkit::minimal_java_version
-    testthat::skip_if(
-        condition = cond_skip_java,
+    testthat::skip_if_not(
+        condition = rjd3jars::check_java_version(silent = TRUE),
         message = "Java version is not sufficient."
     )
     for (k in seq_len(100L)) {
@@ -145,10 +133,8 @@ test_that("rev_set_estimate works", {
 })
 
 test_that("rev_add_usrdefvar works", {
-    cond_skip_java <- rjd3toolkit::get_java_version() <
-        rjd3toolkit::minimal_java_version
-    testthat::skip_if(
-        condition = cond_skip_java,
+    testthat::skip_if_not(
+        condition = rjd3jars::check_java_version(silent = TRUE),
         message = "Java version is not sufficient."
     )
     for (k in seq_len(100L)) {
@@ -169,10 +155,8 @@ test_that("rev_add_usrdefvar works", {
 })
 
 test_that("rev_set_automodel works", {
-    cond_skip_java <- rjd3toolkit::get_java_version() <
-        rjd3toolkit::minimal_java_version
-    testthat::skip_if(
-        condition = cond_skip_java,
+    testthat::skip_if_not(
+        condition = rjd3jars::check_java_version(silent = TRUE),
         message = "Java version is not sufficient."
     )
     for (k in seq_len(100L)) {
@@ -193,10 +177,8 @@ test_that("rev_set_automodel works", {
 })
 
 test_that("rev_set_arima works", {
-    cond_skip_java <- rjd3toolkit::get_java_version() <
-        rjd3toolkit::minimal_java_version
-    testthat::skip_if(
-        condition = cond_skip_java,
+    testthat::skip_if_not(
+        condition = rjd3jars::check_java_version(silent = TRUE),
         message = "Java version is not sufficient."
     )
     for (k in seq_len(100L)) {
@@ -219,10 +201,8 @@ test_that("rev_set_arima works", {
 })
 
 test_that("rev_set_benchmarking works", {
-    cond_skip_java <- rjd3toolkit::get_java_version() <
-        rjd3toolkit::minimal_java_version
-    testthat::skip_if(
-        condition = cond_skip_java,
+    testthat::skip_if_not(
+        condition = rjd3jars::check_java_version(silent = TRUE),
         message = "Java version is not sufficient."
     )
     for (k in seq_len(100L)) {
@@ -243,10 +223,8 @@ test_that("rev_set_benchmarking works", {
 })
 
 test_that("rev_set_outlier works", {
-    cond_skip_java <- rjd3toolkit::get_java_version() <
-        rjd3toolkit::minimal_java_version
-    testthat::skip_if(
-        condition = cond_skip_java,
+    testthat::skip_if_not(
+        condition = rjd3jars::check_java_version(silent = TRUE),
         message = "Java version is not sufficient."
     )
     for (k in seq_len(100L)) {
@@ -267,10 +245,8 @@ test_that("rev_set_outlier works", {
 })
 
 test_that("rev_set_tradingdays works", {
-    cond_skip_java <- rjd3toolkit::get_java_version() <
-        rjd3toolkit::minimal_java_version
-    testthat::skip_if(
-        condition = cond_skip_java,
+    testthat::skip_if_not(
+        condition = rjd3jars::check_java_version(silent = TRUE),
         message = "Java version is not sufficient."
     )
     for (k in seq_len(100L)) {
@@ -291,10 +267,8 @@ test_that("rev_set_tradingdays works", {
 })
 
 test_that("rev_spec works", {
-    cond_skip_java <- rjd3toolkit::get_java_version() <
-        rjd3toolkit::minimal_java_version
-    testthat::skip_if(
-        condition = cond_skip_java,
+    testthat::skip_if_not(
+        condition = rjd3jars::check_java_version(silent = TRUE),
         message = "Java version is not sufficient."
     )
     for (k in seq_len(100L)) {